requirements.txt
requests
python-dotenv
notebook